feat: parent selection

+++ src/patient.rs
@@ -1,10 +1,9 @@
 use serde::Deserialize;
 
-
-
 #[derive(Debug, Deserialize, Clone, Copy)]
 pub struct Patient {
-    pub id: u8,
+    #[serde(default)]
+    pub id: usize,
     pub demand: u32,
     pub start_time: u32,
     pub end_time: u32,
